+++ examples/sample-server/src/resources/pet-photo/index.ts
@@ -1,0 +1,5 @@
+export * from './pet-photo.controller';
+export * from './pet-photo.service';
+export * from './pet-photo.schema';
+export * from './pet-photo.entity';
+export { petPhotoFeature } from './pet-photo.feature';